requirements.txt
flask==2.1.0
flask-cors==3.0.10
boto3==1.18.0
python-dotenv==0.19.0
werkzeug==2.0.1